LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/req.attr/no-referrer/img-tag.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=keep-origin&action=purge&key=583daa9f-d7ff-49f9-819e-1da1657d2546&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=no-redirect&action=purge&key=58f6aa5d-988f-4fe7-b89b-305dc158d947&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=swap-origin&action=purge&key=07d030b6-4309-43bf-93ef-1fdc7aa1cf66&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=keep-origin&action=purge&key=56f814d4-94a0-4518-bfd4-8aaba74b94f7&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=no-redirect&action=purge&key=6ad113f1-6023-4aca-9aaa-c0cf069b4eca&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=swap-origin&action=purge&key=a4aeb4f6-1532-4590-ba56-ee06a5074c4c&path=%2Fmixed-content
CONSOLE MESSAGE: Unable to get image data from canvas. Requested size was 35 x 35
CONSOLE MESSAGE: Unable to get image data from canvas. Requested size was 35 x 35
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?action=purge&key=15bfa6a3-f5e6-47c0-a718-7f5320d682ec&path=%2Fmixed-content


Harness Error (TIMEOUT), message = null

FAIL Referrer Policy: Expects omitted for img-tag to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects omitted for img-tag to same-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "InvalidStateError: The object is in an invalid state."
FAIL Referrer Policy: Expects omitted for img-tag to same-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "InvalidStateError: The object is in an invalid state."
TIMEOUT Referrer Policy: Expects omitted for img-tag to same-http origin and swap-origin redirection from http context. Test timed out
NOTRUN Referrer Policy: Expects omitted for img-tag to same-https origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects omitted for img-tag to same-https origin and no-redirect redirection from http context.
NOTRUN Referrer Policy: Expects omitted for img-tag to same-https origin and swap-origin redirection from http context.
